Switch to `uint64_t` instead of `long long` type
Use of strtoll() makes it impossible to enter values in the range
[LONG_LONG_MAX + 1; ULONG_LONG_MAX].

Also use of signed integers in operations that overflow invokes
undefined behaviour per C standard.

The interface even assumes 64 bits in the type, so I've used uint64_t
instead of `unsigned long long` which can have different width.

%lld and other printf-like formats accept `long long` or `unsigned long
long` and `uint64_t` is likely to be `unsigned long` or some other type
which is different from `unsigned long long` even if its width matches,
hence casts in several places.

+++ include/history.h
@@ -1,6 +1,7 @@
 #ifndef _HISTORY_H
 #define _HISTORY_H
 
+#include <stdint.h>
 
 #define HISTORY_RECORDS_BEFORE_REALLOC 20
 
@@ -18,7 +19,7 @@ extern struct history history;
 
 void clear_history();
 void add_to_history(struct history* h,char* in);
-void add_number_to_history(long long n, int type);
+void add_number_to_history(uint64_t n, int type);
 void browsehistory(char*, int, int*);
 void free_history(struct history *h);
